+++ modules/hdf5/builtin/cpp/Gateway.cpp
@@ -25,6 +25,7 @@
 #include "h5saveBuiltin.hpp"
 #include "h5loadBuiltin.hpp"
 #include "isnh5fileBuiltin.hpp"
+#include "whosnh5Builtin.hpp"
 #include "HDF5_helpers.hpp"
 //=============================================================================
 using namespace Nelson;
@@ -40,6 +41,7 @@ static const nlsGateway gateway[] = {
     { "loadnh5", Nelson::Hdf5Gateway::h5loadBuiltin, 1, 1, CPP_BUILTIN },
     { "savenh5", Nelson::Hdf5Gateway::h5saveBuiltin, 0, 1, CPP_BUILTIN },
     { "isnh5file", Nelson::Hdf5Gateway::isnh5fileBuiltin, 1, 1, CPP_BUILTIN },
+    { "whosnh5", Nelson::Hdf5Gateway::whosnh5Builtin, 1, 1, CPP_BUILTIN },
 };
 //=============================================================================
 static bool